#3397 - Modernize Spring Security API usage
- Move GlobalAuthenticationConfigurerAdapter configuration
