+++ inception/inception-security/src/main/java/de/tudarmstadt/ukp/clarin/webanno/security/config/SecurityAutoConfiguration.java
@@ -32,10 +32,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
 import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.DefaultAuthenticationEventPublisher;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.authentication.configuration.GlobalAuthenticationConfigurerAdapter;
 import org.springframework.security.core.session.SessionRegistry;
 import org.springframework.security.core.session.SessionRegistryImpl;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -111,20 +107,5 @@ public class SecurityAutoConfiguration
         expressionHandler.setApplicationContext(aContext);
         expressionHandler.setPermissionEvaluator(aEvaluator);
         return expressionHandler;
-    }
-
-    @Bean
-    public GlobalAuthenticationConfigurerAdapter globalAuthenticationConfigurerAdapter(
-            AuthenticationProvider authenticationProvider)
-    {
-        return new GlobalAuthenticationConfigurerAdapter()
-        {
-            @Override
-            public void configure(AuthenticationManagerBuilder aAuth) throws Exception
-            {
-                aAuth.authenticationProvider(authenticationProvider);
-                aAuth.authenticationEventPublisher(new DefaultAuthenticationEventPublisher());
-            }
-        };
     }
 }
